fix(hunt): address Copilot #580 review nits
Four Copilot threads on this PR — all valid:

- Tighten the auto-dismiss gate. The bare timestamp check would
  fire on ANY incoming payment after the claim moment (different
  wallet, unrelated Zap, balance-poll catching an unrelated credit).
  Now scoped to active wallet AND, when the wallet detected via the
  expectPayment hash path, also matches the stored paymentHash. The
  hash falls through to walletId-only when detection came via the
  balance-diff fallback (paymentHash is null) — best we can do
  without an invoice hash, same window as before.

- Improve NFC-disabled error UX. Dropping the isNfcEnabled() pre-
  flight (necessary to win the OS-chooser race) left disabled-NFC
  surfacing as a bare 'NFC unavailable on this device' string from
  ensureNfcStarted — inconsistent with NfcWriteSheet / NfcUnlockSheet's
  friendlier 'NFC is turned off' copy. Map the regex match back to
  the same wording at the catch site.

- Extract paymentHashFromBolt11 to src/utils/bolt11.ts as a shared
  helper. ReceiveSheet was already carrying the identical inline
  function and NfcReadSheet had just added a third copy of it. One
  source of truth keeps the dev-warning channel consistent.
